+++ release-notes.md
@@ -16,6 +16,7 @@ See [log](https://github.com/Pathoschild/SMAPI/compare/1.14...1.15).
 
 For players:
 * SMAPI will no longer load mods known to be obsolete or unneeded.
+* When the `ObjectInformation.xnb` is broken, SMAPI will now print one error to the console instead of a warning flood. (The individual issues are still listed in the log file if needed.)
 
 For modders:
 * You can now specify minimum dependency versions in `manifest.json`.